refactor: use list hoc to reduce code
